Move TaxRate and Leverage classes into transformation package
