+++ project/modules/performance/analyzers.py
@@ -7,7 +7,7 @@ from typing import Dict
 
 import pandas as pd
 
-from .transformations import TaxRate, Leverage
+from .transformation import TaxRate, Leverage
 from .annualizer import Annualizer
 from .metrics import MetricsCollection, SpearmanCorrelation
