+++ project/modules/performance/transformation/tax_rate.py
@@ -1,4 +1,4 @@
-"""税率・レバレッジ計算用クラス群"""
+"""税金の計算と適用を管理するモジュール"""
 
 from __future__ import annotations
 
@@ -22,18 +22,3 @@ class TaxRate:
         untaxed = returns.copy()
         untaxed = untaxed / (1 - self.tax_rate)
         return untaxed
-
-
-class Leverage:
-    """レバレッジの適用と解除を管理する。"""
-
-    def __init__(self, leverage_ratio: float = 3.1) -> None:
-        self.leverage_ratio = leverage_ratio
-
-    def apply_leverage(self, returns: pd.Series) -> pd.Series:
-        """レバレッジ適用後リターンを計算する。"""
-        return returns * self.leverage_ratio
-
-    def remove_leverage(self, leveraged_returns: pd.Series) -> pd.Series:
-        """レバレッジ除去後のリターンを計算する。"""
-        return leveraged_returns / self.leverage_ratio
